pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Groq-powered RAG chatbot"
authors = [{name = "ChatGPT"}]
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.112",
  "uvicorn>=0.30",
  "langchain>=0.2.8",
  "langchain-community>=0.2.8",
  "llama-index-core>=0.14",
  "llama-index-embeddings-huggingface>=0.6",
  "llama-index-llms-groq>=0.4",
  "llama-index-vector-stores-qdrant>=0.8",
  "qdrant-client>=1.10",
  "pydantic>=2.8",
  "pydantic-settings>=2.3",
  "python-dotenv>=1.0",
  "docx2txt>=0.8",
  "python-docx>=1.1",
  "pypdf>=4.2",
  "python-multipart>=0.0.20",
  "httpx>=0.27",
  "streamlit>=1.40",
  "pytest>=8.3"
]
